// store/search/types/enums/action-type.ts
export enum ActionType {
    SET_QUERY = "SET_QUERY",

    SEARCH_LIVE = "SEARCH_LIVE",
    LOAD_LIVE_NEXT_PAGE = "LOAD_LIVE_NEXT_PAGE",

    SEARCH_MOVIES = "SEARCH_MOVIES",
    LOAD_MOVIES_NEXT_PAGE = "LOAD_MOVIES_NEXT_PAGE",

    SEARCH_SHOWS = "SEARCH_SHOWS",
    LOAD_SHOWS_NEXT_PAGE = "LOAD_SHOWS_NEXT_PAGE",

    SEARCH_CHANNELS = "SEARCH_CHANNELS",
    LOAD_CHANNELS_NEXT_PAGE = "LOAD_CHANNELS_NEXT_PAGE",

    SEARCH_CREDITS = "SEARCH_CREDITS",
    LOAD_CREDITS_NEXT_PAGE = "LOAD_CREDITS_NEXT_PAGE",
}